KINSI_Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
email-validator==2.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
